refactor dashboard tabs component
