+++ packages/backend/src/models/DashboardModel/DashboardModel.ts
@@ -136,7 +136,6 @@ export class DashboardModel {
                 name: tab.name,
                 uuid: tab.uuid,
                 dashboard_id: dashboardId,
-                is_default: tab.isDefault,
                 order: tab.order,
             });
         });
@@ -593,7 +592,6 @@ export class DashboardModel {
             .select<DashboardTab[]>(
                 `${DashboardTabsTableName}.name`,
                 `${DashboardTabsTableName}.uuid`,
-                `${DashboardTabsTableName}.is_default`,
                 `${DashboardTabsTableName}.order`,
             )
             .where(
